Add "sip-status-file" option
This passes a file parameter to t38modem where the SIP connection
status code will be written. This is useful for clients that want to
monitor the SIP status codes during a call and may choose to restart the
t38modem application instead of letting opal attempt a retry.

The string format is: {status:<STATUS_CODE>,remote:<REMOTE_ADDRESS>}